Create MachineLearningProject
Machine learning is the field of study that gives computers the capability to learn without being explicitly programmed. It is a subset of Artificial Intelligence (AI). Basically, machine learning is giving enough data to the computer/model to create or find patterns in the data which it will use when additional data is provided to get the results. This ability to learn without writing codes or program makes the machine act like human. Nowadays, it is being used almost in every field including self-driving cars, robotics, language processing, vision processing, forecasting stock market trends etc.